pyproject.toml
[tool.poetry]
name = "electronics-shop-project"
version = "0.1.0"
description = ""
authors = ["kirillskyeng <kirill.netreba@skyeng.ru>"]
readme = "README.md"
packages = [{include = "electronics_shop_project"}]

[tool.poetry.dependencies]
python = "3.10.6"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
